Remove "ReadOnly" component
